+++ dashboard/state_aggregator.py
@@ -362,10 +362,10 @@ class StateAPIManager:
         """
         job_id = None
         for filter in option.filters:
-            if filter[0] == "job_id":
+            if filter[0] == "job_id" and filter[1] == "=":
+                # Filtering by job_id == xxxx, pass it to source side filtering.
                 # tuple consists of (job_id, predicate, value)
                 job_id = filter[2]
-
         try:
             reply = await self._client.get_all_task_info(
                 timeout=option.timeout,
